+++ packages/adp-tooling/src/prompts/index.ts
@@ -1,3 +1,5 @@
 export { getPrompts as getPromptsForChangeDataSource } from './change-data-source';
+export { getPrompts as getPromptsForAddComponentUsages } from './add-component-usages';
+export { getPrompts as getPromptsForNewModel } from './add-new-model';
 export { getPrompts as getPromptsForChangeInbound } from './change-inbound';
 export { getPrompts as getPromptsForAddAnnotationsToOData } from './add-annotations-to-odata';