+++ packages/i18n/src/parser/utils/index.ts
@@ -1,4 +1,0 @@
-export { getLineOffsets } from './line-offsets';
-export { positionAt } from './position';
-export { rangeAt } from './range';
-export { Position, Range } from './language-server';