+++ packages/ui5-info/src/ui5-version-fallback.ts
@@ -8,11 +8,27 @@ export const supportState = {
     skipped: 'Skipped'
 } as const;
 
-// Updated April-16-2024 from https://ui5.sap.com/versionoverview.json
+// Updated Aug-08-2024 from https://ui5.sap.com/versionoverview.json
 export const ui5VersionFallbacks = [
     {
-        version: '1.122.*',
+        version: '1.126.*',
         support: supportState.maintenance
+    },
+    {
+        version: '1.125.*',
+        support: supportState.outOfMaintenance
+    },
+    {
+        version: '1.124.*',
+        support: supportState.maintenance
+    },
+    {
+        version: '1.123.*',
+        support: supportState.outOfMaintenance
+    },
+    {
+        version: '1.122.*',
+        support: supportState.outOfMaintenance
     },
     {
         version: '1.121.*',
@@ -32,7 +48,7 @@ export const ui5VersionFallbacks = [
     },
     {
         version: '1.117.*',
-        support: supportState.maintenance
+        support: supportState.outOfMaintenance
     },
     {
         version: '1.116.*',